+++ src/components/SignUp/index.jsx
@@ -1,5 +1,5 @@
 import "./styles.css";
-
+import { useState } from "react";
 export default function SignUp() {
   const [inputs, setInputs] = useState({});
